requirements.txt
asn1crypto==0.24.0
attrs==17.4.0
blessed==1.15.0
botocore==1.12.111
cached-property==1.5.1
cachetools==3.1.0
cement==2.8.2
certifi==2019.3.9
chardet==3.0.4
Click==7.0
colorama==0.4.1
cryptography==2.1.4
cycler==0.10.0
cymem==2.0.2
cytoolz==0.9.0.1
decorator==4.1.2
dill==0.2.9
docker==3.7.0
docker-compose==1.23.2
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
dominate==2.3.5
evdev==0.7.0
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
future==0.16.0
google-api-python-client==1.7.8
google-auth==1.6.2
google-auth-httplib2==0.0.3
gunicorn==19.9.0
httplib2==0.12.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
jmespath==0.9.4
joblib==0.13.1
jsonschema==2.6.0
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.1.1
matplotlib==3.0.2
msgpack==0.5.6
msgpack-numpy==0.4.3.2
murmurhash==1.0.1
netifaces==0.10.4
nose==1.3.7
numpy==1.16.1
oauth==1.0.1
oauth2client==4.1.3
olefile==0.45.1
pandas==0.24.1
pathspec==0.5.9
pexpect==4.2.1
Pillow==5.1.0
plac==0.9.6
pluggy==0.6.0
preshed==2.0.1
protobuf==3.0.0
py==1.5.2
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycairo==1.16.2
pycodestyle==2.5.0
pycrypto==2.6.1
pycups==1.9.73
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.3.1
pyRFC3339==1.0
pytest==3.3.2
python-dateutil==2.8.0
python-debian==0.1.32
pytz==2018.9
pyudev==0.21.0
pyxdg==0.25
regex==2018.1.10
reportlab==3.4.0
requests==2.21.0
requests-unixsocket==0.1.5
rsa==4.0
s3transfer==0.2.0
scikit-learn==0.20.2
scipy==1.2.0
seaborn==0.9.0
SecretStorage==2.3.1
semantic-version==2.5.0
simplejson==3.13.2
six==1.12.0
sklearn==0.0
soupsieve==1.7.3
spacy-nightly
systemd-python==234
termcolor==1.1.0
texttable==0.9.1
thinc==6.12.1
toolz==0.9.0
tqdm==4.31.1
ujson==1.35
uritemplate==3.0.0
urllib3==1.24.1
visitor==0.1.3
wadllib==1.3.2
wcwidth==0.1.7
websocket-client==0.55.0
Werkzeug==0.14.1
wrapt==1.10.11
WTForms==2.2.1
zope.interface==4.3.2
